optimally clean entry point
