+++ lib/meta.js
@@ -1,0 +1,6 @@
+'use strict'
+
+const { version } = require('../package.json')
+const LOG_VERSION = 1
+
+module.exports = { version, LOG_VERSION }
