+++ lib/tools.js
@@ -3,6 +3,7 @@
 const format = require('quick-format-unescaped')
 const util = require('util')
 const serializers = require('pino-std-serializers')
+const SonicBoom = require('sonic-boom')
 const {
   lsCacheSym,
   chindingsSym,
@@ -201,10 +202,35 @@ function asChindings (instance, bindings) {
   return data
 }
 
+function createArgsNormalizer (defaultOptions) {
+  return function normalizeArgs (opts = {}, stream) {
+    if (opts instanceof SonicBoom || opts.writable || opts._writableState) {
+      stream = opts
+      opts = null
+    }
+    opts = Object.assign({}, defaultOptions, opts)
+    if ('extreme' in opts) {
+      throw new Error('The extreme option has been removed, use require(\'pino\').extreme(dest) instead')
+    }
+    const { enabled, prettyPrint, prettifier, messageKey } = opts
+    if (enabled === false) opts.level = 'silent'
+    stream = stream || process.stdout
+    if (stream === process.stdout && stream.fd >= 0) {
+      stream = new SonicBoom(stream.fd)
+    }
+    if (prettyPrint) {
+      const prettyOpts = Object.assign({ messageKey }, prettyPrint)
+      stream = getPrettyStream(prettyOpts, prettifier, stream)
+    }
+    return { opts, stream }
+  }
+}
+
 module.exports = {
   noop,
   getPrettyStream,
   asChindings,
   asJson,
-  genLog
+  genLog,
+  createArgsNormalizer
 }
